chore: Separate Routine API into Individual Files
